# api_gateway/docs/swagger.yaml
basePath: /
definitions:
  entity.AddChannelReq:
    properties:
      channel_name:
        type: string
    type: object
  entity.AddChannelReqChannel:
    properties:
      channel_name:
        type: string
    type: object
  entity.AddSpeakerReq:
    properties:
      model_name:
        type: string
    type: object
  entity.AddTVReq:
    properties:
      model_name:
        type: string
    type: object
  entity.Channel:
    properties:
      channel_name:
        type: string
      channel_number:
        type: string
    type: object
...
  entity.DeleteChannelReq:
    properties:
      channel_name:
        type: string
    type: object
...
  entity.DownOrUpVoiceTvRes:
    properties:
      sound:
        type: integer
    type: object
  entity.GetAllUserRes:
    properties:
      count:
        type: integer
      users:
        items:
          $ref: '#/definitions/entity.User'
        type: array
    type: object
...
  entity.PreviousAndNextRes:
    properties:
      channel:
        $ref: '#/definitions/entity.Channel'
    type: object
  entity.Profile:
    properties:
      address:
        type: string
      created_at:
        type: string
      deleted_at:
        type: string
      first_name:
        type: string
      updated_at:
        type: string
    type: object
  entity.StatusMessage:
    properties:
      message:
        type: string
    type: object
...
  entity.User:
    properties:
      email:
        type: string
      first_name:
        type: string
      id:
        type: string
      last_name:
        type: string
      password:
        type: string
      profile:
        $ref: '#/definitions/entity.Profile'
    type: object
...
host: localhost:9002
info:
  contact: {}
  description: This is a sample server for a restaurant reservation system.
  title: Artisan Connect
  version: "1.0"
paths:
  /speaker/channel:
    delete:
      consumes:
      - application/json
      description: Add tv to home
      parameters:
      - description: channel create information
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/entity.DeleteChannelReq'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/entity.StatusMessage'
        "400":
          description: Bad Request
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
      security:
      - ApiKeyAuth: []
      - BearerAuth: []
      summary: DeleteChannel
      tags:
      - television
    post:
      consumes:
      - application/json
      description: Add tv to home
      parameters:
      - description: channel create information
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/entity.AddChannelReqChannel'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/entity.StatusMessage'
        "400":
          description: Bad Request
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
      security:
      - ApiKeyAuth: []
      - BearerAuth: []
      summary: AddChannel
      tags:
      - television
  /speaker/channel/cursor:
    post:
      consumes:
      - application/json
      description: Add tv to home
      parameters:
      - description: true or false
        in: query
        name: next
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/entity.PreviousAndNextRes'
        "400":
          description: Bad Request
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
      security:
      - ApiKeyAuth: []
      - BearerAuth: []
      summary: PreviousAndNext
      tags:
      - television
  /speaker/open:
    put:
      consumes:
      - application/json
      description: Add tv to home
      parameters:
      - description: true or false
        in: query
        name: open
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/entity.StatusMessage'
        "400":
          description: Bad Request
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
      security:
      - ApiKeyAuth: []
      - BearerAuth: []
      summary: OpenSpeaker
      tags:
      - television
  /speaker/register:
    post:
      consumes:
      - application/json
      description: AddSpeaker add speaker
      parameters:
      - description: tv create information
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/entity.AddSpeakerReq'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/entity.StatusMessage'
        "400":
          description: Bad Request
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
      security:
      - ApiKeyAuth: []
      - BearerAuth: []
      summary: AddSpeaker
      tags:
      - speaker
  /speaker/user/channels:
    get:
      consumes:
      - application/json
      description: Add tv to home
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/entity.GetAllUserRes'
        "400":
          description: Bad Request
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
      security:
      - ApiKeyAuth: []
      - BearerAuth: []
      summary: GetChannels
      tags:
      - television
  /speaker/voice:
    post:
      consumes:
      - application/json
      description: Add tv to home
      parameters:
      - description: true or false
        in: query
        name: up
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/entity.DownOrUpVoiceTvRes'
        "400":
          description: Bad Request
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
      security:
      - ApiKeyAuth: []
      - BearerAuth: []
      summary: ControlVoice
      tags:
      - television
  /tv/channel:
    delete:
      consumes:
      - application/json
      description: Add tv to home
      parameters:
      - description: channel create information
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/entity.DeleteChannelReq'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/entity.StatusMessage'
        "400":
          description: Bad Request
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
      security:
      - ApiKeyAuth: []
      - BearerAuth: []
      summary: DeleteChannel
      tags:
      - television
    post:
      consumes:
      - application/json
      description: Add tv to home
      parameters:
      - description: channel create information
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/entity.AddChannelReq'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/entity.StatusMessage'
        "400":
          description: Bad Request
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
      security:
      - ApiKeyAuth: []
      - BearerAuth: []
      summary: AddChannel
      tags:
      - television
  /tv/channel/cursor:
    post:
      consumes:
      - application/json
      description: Add tv to home
      parameters:
      - description: true or false
        in: query
        name: next
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/entity.PreviousAndNextRes'
        "400":
          description: Bad Request
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
      security:
      - ApiKeyAuth: []
      - BearerAuth: []
      summary: PreviousAndNext
      tags:
      - television
  /tv/open:
    put:
      consumes:
      - application/json
      description: Add tv to home
      parameters:
      - description: true or false
        in: query
        name: open
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/entity.StatusMessage'
        "400":
          description: Bad Request
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
      security:
      - ApiKeyAuth: []
      - BearerAuth: []
      summary: OpenTv
      tags:
      - television
  /tv/register:
    post:
      consumes:
      - application/json
      description: Add tv to home
      parameters:
      - description: tv create information
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/entity.AddTVReq'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/entity.StatusMessage'
        "400":
          description: Bad Request
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
      security:
      - ApiKeyAuth: []
      - BearerAuth: []
      summary: AddTvHome
      tags:
      - television
  /tv/user/channels:
    get:
      consumes:
      - application/json
      description: Add tv to home
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/entity.GetAllUserRes'
        "400":
          description: Bad Request
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
      security:
      - ApiKeyAuth: []
      - BearerAuth: []
      summary: GetChannels
      tags:
      - television
  /tv/voice:
    post:
      consumes:
      - application/json
      description: Add tv to home
      parameters:
      - description: true or false
        in: query
        name: up
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/entity.DownOrUpVoiceTvRes'
        "400":
          description: Bad Request
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
      security:
      - ApiKeyAuth: []
      - BearerAuth: []
      summary: ControlVoice
      tags:
      - television
...
schemes:
- http
securityDefinitions:
  ApiKeyAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"